CMNeXt/requirements.txt
# CMNeXt UAVScenes Requirements
# Python 3.8, CUDA 11.3

# PyTorch - MUST install first, pinned to prevent override
torch==1.11.0+cu113
torchvision==0.12.0+cu113
--extra-index-url https://download.pytorch.org/whl/cu113

# Core
numpy>=1.21.0,<2.0

# Image processing
opencv-python>=4.5.0
Pillow>=8.0.0

# Config & logging
PyYAML>=5.4
tensorboard>=2.10.0

# Utilities
tqdm>=4.60.0
matplotlib>=3.5.0
scipy>=1.7.0

# timm - pinned to work with torch 1.11
timm==0.6.13
